refactored unittests
