+++ flygym/data/pose/pose_tripod.yaml
