--- flygym/tests/__init__.py
@@ -1,4 +1,0 @@
-import tempfile
-from pathlib import Path
-
-temp_base_dir = Path(tempfile.gettempdir()) / "flygym_test"
